requirements.txt
playwright
streamlit
chromadb
sentence-transformers
google-generativeai
SpeechRecognition
pyttsx3
python-dotenv 